Update: Move bard fact adapter to client
