+++ packages/client/package.json
@@ -66,6 +66,8 @@
   },
   "dependencies": {
     "core-js": "^3.22.8",
+    "cross-fetch": "^3.1.5",
+    "effection": "^2.0.4",
     "lodash": "^4.17.21",
     "moment": "^2.29.3",
     "tiny-invariant": "^1.2.0"
