+++ packages/client/src/services/interfaces/registry.ts
@@ -1,0 +1,7 @@
+/**
+ * Copyright 2022, Yahoo Holdings Inc.
+ * Licensed under the terms of the MIT license. See accompanying LICENSE.md file for terms.
+ */
+
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+export default interface ServiceRegistry {}
