+++ packages/data/package.json
@@ -31,6 +31,7 @@
     "postpublish": "ember ts:clean"
   },
   "dependencies": {
+    "@ember/test-waiters": "^3.0.2",
     "@faker-js/faker": "^6.0.0",
     "@miragejs/graphql": "^0.1.13",
     "@rsql/parser": "^1.0.2",
@@ -63,8 +64,8 @@
     "@types/ember": "^3.16.5",
     "@types/ember__array": "^3.16.5",
     "@types/ember__routing": "^3.16.16",
-    "@types/ember-data__serializer": "^3.16.2",
     "@types/ember__test-helpers": "^1.7.6",
+    "@types/ember-data__serializer": "^3.16.2",
     "@types/ember-qunit": "^3.4.15",
     "@types/faker": "^5.1.6",
     "@types/lodash-es": "^4.17.3",
